# api/collaboration/openapi.yaml
openapi: 3.1.0
info:
  title: AphelionDMM Collaboration Control API
  version: 1.0.0
  description: Version 1 lifecycle and snapshot contract. Filesystem paths and process commands are never accepted.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1
  - url: https://collaboration.example
paths:
  /v1/auth/begin:
    post:
      summary: Begin hosted OIDC authentication
      operationId: beginHostedAuthentication
      security: []
      responses:
        '200':
          description: OIDC authorization URL containing the one-use state value
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedAuthBegin' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/auth/desktop/begin:
    post:
      summary: Begin verifier-bound desktop OIDC authentication
      operationId: beginHostedDesktopAuthentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HostedDesktopAuthBeginRequest' }
      responses:
        '200':
          description: OIDC authorization URL and opaque desktop handoff identifier
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedDesktopAuthBegin' }
        '400': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/auth/desktop/exchange:
    post:
      summary: Exchange a completed desktop handoff for an in-memory hosted credential
      operationId: exchangeHostedDesktopAuthentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HostedDesktopAuthExchangeRequest' }
      responses:
        '200':
          description: Short-lived hosted authentication session
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedAuthentication' }
        '202':
          description: Browser authentication has not completed yet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedDesktopAuthPending' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/complete:
    get:
      summary: Complete hosted OIDC authentication
      operationId: completeHostedAuthentication
      security: []
      parameters:
        - { name: state, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 512 } }
        - { name: code, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 4096 } }
      responses:
        '200':
          description: Short-lived hosted authentication session
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/HostedAuthentication' }
                  - { $ref: '#/components/schemas/HostedDesktopAuthComplete' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/logout:
    post:
      summary: End a hosted authentication session
      operationId: logoutHostedAuthentication
      security: [{ bearerToken: [] }]
      responses:
        '204': { description: Hosted authentication session removed }
        '503': { $ref: '#/components/responses/Error' }
  /v1/hosted/sessions:
    post:
      summary: Create a durable hosted collaboration session
      operationId: createHostedSession
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
      responses:
        '201':
          description: Durable hosted session created with the authenticated identity as owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedSessionResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
  /v1/sessions:
    post:
      summary: Create an embedded collaboration session
      operationId: createSession
      security: [{ launchToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateSessionResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    get:
      summary: Get collaboration session metadata
      operationId: getSession
      security: [{ bearerToken: [] }]
      responses:
        '200':
          description: Session metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}/join-tokens:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Create a short-lived embedded-mode join token
      description: Hosted sessions use OIDC-bound hosted invitations and cannot mint embedded identities.
      operationId: createJoinToken
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateJoinTokenRequest' }
      responses:
        '201':
          description: Short-lived scoped join token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JoinToken' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}/hosted-invitations:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Create a one-use hosted invitation
      operationId: createHostedInvitation
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateHostedInvitationRequest' }
      responses:
        '201':
          description: One-use hosted invitation. The raw token is returned only in this response.
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedInvitation' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}/hosted-invitations/redeem:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Redeem a one-use hosted invitation
      operationId: redeemHostedInvitation
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RedeemHostedInvitationRequest' }
      responses:
        '200':
          description: Hosted membership created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedMembership' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}/snapshot:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    get:
      summary: Get the authoritative collaboration snapshot
      operationId: getSnapshot
      security: [{ bearerToken: [] }]
      responses:
        '200':
          description: Authoritative collaboration snapshot
          headers:
            ETag:
              schema: { type: string, maxLength: 128 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snapshot' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}/exports:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Request a staged export checkpoint
      operationId: createExportCheckpoint
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExportRequest' }
      responses:
        '202':
          description: Export checkpoint accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExportCheckpoint' }
        '403': { $ref: '#/components/responses/Error' }
        '501': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    launchToken: { type: http, scheme: bearer, description: Single-use embedded launch token }
    bearerToken: { type: http, scheme: bearer }
...
  responses:
    Error:
      description: Bounded stable error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    SHA256:
      type: string
      pattern: '^[0-9a-f]{64}$'
...
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [snapshot]
      properties:
        snapshot: { $ref: '#/components/schemas/Snapshot' }
        display_name: { type: string, minLength: 1, maxLength: 128, default: Owner }
    CreateSessionResponse:
      type: object
      additionalProperties: false
      required: [session_id, document_id, revision, map_hash, owner_token, owner_token_expires_at]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        document_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 0 }
        map_hash: { $ref: '#/components/schemas/SHA256' }
        owner_token: { type: string, minLength: 1, maxLength: 128 }
        owner_token_expires_at: { type: string, format: date-time }
    HostedSessionResponse:
      type: object
      additionalProperties: false
      required: [session_id, document_id, revision, map_hash]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        document_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 0 }
        map_hash: { $ref: '#/components/schemas/SHA256' }
    HostedAuthBegin:
      type: object
      additionalProperties: false
      required: [authorization_url]
      properties:
        authorization_url: { type: string, format: uri, maxLength: 8192 }
    HostedDesktopAuthBeginRequest:
      type: object
      additionalProperties: false
      required: [verifier_challenge]
      properties:
        verifier_challenge: { type: string, minLength: 43, maxLength: 43, pattern: '^[A-Za-z0-9_-]{43}$' }
    HostedDesktopAuthBegin:
      type: object
      additionalProperties: false
      required: [authorization_url, handoff_id]
      properties:
        authorization_url: { type: string, format: uri, maxLength: 8192 }
        handoff_id: { type: string, minLength: 1, maxLength: 128 }
    HostedDesktopAuthExchangeRequest:
      type: object
      additionalProperties: false
      required: [handoff_id, verifier]
      properties:
        handoff_id: { type: string, minLength: 1, maxLength: 128 }
        verifier: { type: string, minLength: 43, maxLength: 256 }
    HostedDesktopAuthPending:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, const: pending }
    HostedDesktopAuthComplete:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, const: complete }
    HostedAuthentication:
      type: object
      additionalProperties: false
      required: [token, actor_id, display_name, expires_at]
      properties:
        token: { type: string, minLength: 1, maxLength: 128 }
        actor_id: { type: string, format: uuid }
        display_name: { type: string, minLength: 1, maxLength: 128 }
        expires_at: { type: string, format: date-time }
...